Adapt the Gfal2 wrapper to return compatible integer types with the Gfal2 transfer parameter functions

+++ src/url-copy/Gfal2.h
@@ -178,10 +178,10 @@ public:
         }
     }
 
-    unsigned getNumberOfStreams(void)
+    uint32_t getNumberOfStreams(void)
     {
         GError *error = NULL;
-        unsigned val = gfalt_get_nbstreams(params, &error);
+        uint32_t val = gfalt_get_nbstreams(params, &error);
         if (error) {
             throw Gfal2Exception(error);
         }
@@ -196,20 +196,20 @@ public:
         }
     }
 
-    unsigned getTcpBuffersize(void)
+    uint64_t getTcpBuffersize(void)
     {
         GError *error = NULL;
-        unsigned val = gfalt_get_tcp_buffer_size(params, &error);
+        uint64_t val = gfalt_get_tcp_buffer_size(params, &error);
         if (error) {
             throw Gfal2Exception(error);
         }
         return val;
     }
 
-    unsigned getTimeout(void)
+    uint64_t getTimeout(void)
     {
         GError *error = NULL;
-        unsigned val = gfalt_get_timeout(params, &error);
+        uint64_t val = gfalt_get_timeout(params, &error);
         if (error) {
             throw Gfal2Exception(error);
         }
